refactor interfaces for response of evaluation results endpoints
