+++ src/evaluation/service/evaluation.service.ts
@@ -8,16 +8,16 @@ import { SimulationRepository } from '@db/repositories/simulation.repository';
 
 import { RunEvaluationRequest } from '@evaluation/model/request/run-evaluation.request';
 import { RunEvaluationResponse } from '@evaluation/model/response/run-evaluation.response';
-import {
-  EvaluationExecuted,
-  EvaluationResultForConversation,
-  EvaluationResultForSimulation,
-} from '@evaluation/model/response/evaluation-result.response';
+
 import metricService from '@evaluation/service/metric.service';
 
 import { EvaluationStatus } from '@enums/evaluation-status.enum';
 
 import simulationService from '@simulation/service/simulation.service';
+import EvaluationResultForConversation, {
+  EvaluationExecuted,
+} from '@evaluation/model/response/results-for-conversation.response';
+import EvaluationResultForSimulation from '@evaluation/model/response/results-for-simulation.response';
 
 const evaluationRepository = new EvaluationRepository(EvaluationModel);
 const conversationRepository = new ConversationRepository(ConversationModel);
